Move logic into a separate mapper
