+++ src/maps/dashboard.ts
@@ -1,52 +1,9 @@
-import { DateRangeSelectRangeValue, DateRangeSelectSpanValue, DateRangeSelectValue } from '@prefecthq/prefect-design'
-import { addSeconds, differenceInSeconds } from 'date-fns'
 import { FlowRunsFilter, TaskRunsFilter, TaskRunsHistoryFilter, WorkPoolWorkersFilter } from '@/models/Filters'
 import { MapFunction } from '@/services/Mapper'
 import { WorkspaceDashboardFilter } from '@/types/dashboard'
-import { sortDates } from '@/utilities/dates'
-
-function nowWithoutMilliseconds(): Date {
-  const now = new Date()
-
-  now.setMilliseconds(0)
-
-  return now
-}
-
-type DateRange = {
-  startDate: Date,
-  endDate: Date,
-  timeSpanInSeconds: number,
-}
-
-function mapDateRangeSelectRangeValue({ startDate, endDate }: DateRangeSelectRangeValue): DateRange {
-  const timeSpanInSeconds = differenceInSeconds(endDate, startDate)
-
-  return { startDate, endDate, timeSpanInSeconds }
-}
-
-function mapDateRangeSelectSpanValue({ seconds }: DateRangeSelectSpanValue): DateRange {
-  const now = nowWithoutMilliseconds()
-  const then = addSeconds(now, seconds)
-  const [startDate, endDate] = [now, then].sort(sortDates)
-
-  return { startDate, endDate, timeSpanInSeconds: seconds }
-}
-
-function mapDateRangeSelectValue(value: NonNullable<DateRangeSelectValue>): DateRange {
-  switch (value.type) {
-    case 'range':
-      return mapDateRangeSelectRangeValue(value)
-    case 'span':
-      return mapDateRangeSelectSpanValue(value)
-    default:
-      const exhaustive: never = value
-      throw new Error(`No handler for DateRangeSelectValue.type: ${exhaustive}`)
-  }
-}
 
 export const mapWorkspaceDashboardFilterToTaskRunsFilter: MapFunction<WorkspaceDashboardFilter, TaskRunsFilter> = function(source) {
-  const { startDate, endDate } = mapDateRangeSelectValue(source.range)
+  const { startDate, endDate } = this.map('DateRangeSelectValue', source.range, 'DateRange')
 
   return {
     flowRuns: {
@@ -62,7 +19,7 @@ export const mapWorkspaceDashboardFilterToTaskRunsFilter: MapFunction<WorkspaceD
 }
 
 export const mapWorkspaceDashboardFilterToTaskRunsHistoryFilter: MapFunction<WorkspaceDashboardFilter, TaskRunsHistoryFilter> = function(source) {
-  const { startDate, endDate, timeSpanInSeconds } = mapDateRangeSelectValue(source.range)
+  const { startDate, endDate, timeSpanInSeconds } = this.map('DateRangeSelectValue', source.range, 'DateRange')
 
   return {
     historyStart: startDate,
@@ -77,7 +34,7 @@ export const mapWorkspaceDashboardFilterToTaskRunsHistoryFilter: MapFunction<Wor
 }
 
 export const mapWorkspaceDashboardFilterToFlowRunsFilter: MapFunction<WorkspaceDashboardFilter, FlowRunsFilter> = function(source) {
-  const { startDate, endDate } = mapDateRangeSelectValue(source.range)
+  const { startDate, endDate } = this.map('DateRangeSelectValue', source.range, 'DateRange')
 
   const filter: FlowRunsFilter = {
     flowRuns: {
@@ -93,7 +50,7 @@ export const mapWorkspaceDashboardFilterToFlowRunsFilter: MapFunction<WorkspaceD
 }
 
 export const mapWorkspaceDashboardFilterToWorkPoolWorkersFilter: MapFunction<WorkspaceDashboardFilter, WorkPoolWorkersFilter> = function(source) {
-  const { startDate, endDate } = mapDateRangeSelectValue(source.range)
+  const { startDate, endDate } = this.map('DateRangeSelectValue', source.range, 'DateRange')
 
   return {
     workers: {
